requirements.txt
pygame
edge-tts
requests
python-dotenv
bilibili-api-python
